+++ R/form.R
@@ -6,29 +6,29 @@
 #'   \url{http://www.w3.org/TR/html401/interact/forms.html}
 #' @examples
 #' \donttest{
-#' html_form(html("https://hadley.wufoo.com/forms/libraryrequire-quiz/"))
-#' html_form(html("https://hadley.wufoo.com/forms/r-journal-submission/"))
+#' html_form(read_html("https://hadley.wufoo.com/forms/libraryrequire-quiz/"))
+#' html_form(read_html("https://hadley.wufoo.com/forms/r-journal-submission/"))
 #'
-#' box_office <- html("http://www.boxofficemojo.com/movies/?id=ateam.htm")
+#' box_office <- read_html("http://www.boxofficemojo.com/movies/?id=ateam.htm")
 #' box_office %>% html_node("form") %>% html_form()
 #' }
 html_form <- function(x) UseMethod("html_form")
 
 #' @export
-html_form.XMLAbstractDocument <- function(x) {
-  html_form(html_nodes(x, "form"))
+html_form.xml_document <- function(x) {
+  html_form(xml2::xml_find_all(x, ".//form"))
 }
 
 #' @export
-html_form.XMLNodeSet <- function(x) {
+html_form.xml_nodeset <- function(x) {
   lapply(x, html_form)
 }
 
 #' @export
-html_form.XMLInternalElementNode <- function(x) {
-  stopifnot(inherits(x, "XMLAbstractNode"), XML::xmlName(x) == "form")
+html_form.xml_node <- function(x) {
+  stopifnot(xml2::xml_name(x) == "form")
 
-  attr <- as.list(XML::xmlAttrs(x))
+  attr <- as.list(xml2::xml_attrs(x))
   name <- attr$id %||% attr$name %||% "<unnamed>" # for human readers
   method <- toupper(attr$method) %||% "GET"
   enctype <- convert_enctype(attr$enctype)
@@ -76,7 +76,7 @@ parse_fields <- function(form) {
   raw <- html_nodes(form, "input, select, textarea, button")
 
   fields <- lapply(raw, function(x) {
-    switch(XML::xmlName(x),
+    switch(xml2::xml_name(x),
       textarea = parse_textarea(x),
       input = parse_input(x),
       select = parse_select(x),
@@ -109,8 +109,8 @@ print.fields <- function(x, ..., indent = 0) {
 # * date/datetime/month/week/time
 # * (if unknown treat as text)
 parse_input <- function(input) {
-  stopifnot(inherits(input, "XMLAbstractNode"), XML::xmlName(input) == "input")
-  attr <- as.list(XML::xmlAttrs(input))
+  stopifnot(inherits(input, "xml_node"), xml2::xml_name(input) == "input")
+  attr <- as.list(xml2::xml_attrs(input))
 
   structure(
     list(
@@ -127,9 +127,9 @@ parse_input <- function(input) {
 }
 
 parse_select <- function(select) {
-  stopifnot(inherits(select, "XMLAbstractNode"), XML::xmlName(select) == "select")
+  stopifnot(inherits(select, "xml_node"), xml2::xml_name(select) == "select")
 
-  attr <- as.list(XML::xmlAttrs(select))
+  attr <- as.list(xml2::xml_attrs(select))
   options <- parse_options(html_nodes(select, "option"))
 
   structure(
@@ -149,10 +149,10 @@ format.select <- function(x, ...) {
 
 parse_options <- function(options) {
   parse_option <- function(option) {
-    attr <- as.list(XML::xmlAttrs(option))
+    attr <- as.list(xml2::xml_attrs(option))
     list(
       value = attr$value,
-      name = XML::xmlValue(option),
+      name = xml2::xml_text(option),
       selected = !is.null(attr$selected)
     )
   }
@@ -169,12 +169,12 @@ parse_options <- function(options) {
 }
 
 parse_textarea <- function(textarea) {
-  attr <- as.list(XML::xmlAttrs(textarea))
+  attr <- as.list(xml2::xml_attrs(textarea))
 
   structure(
     list(
       name = attr$name,
-      value = XML::xmlValue(textarea)
+      value = xml2::xml_text(textarea)
     ),
     class = "textarea"
   )
@@ -186,8 +186,8 @@ format.textarea <- function(x, ...) {
 }
 
 parse_button <- function(button) {
-  stopifnot(inherits(button, "XMLAbstractNode"), XML::xmlName(button) == "button")
-  attr <- as.list(XML::xmlAttrs(button))
+  stopifnot(inherits(button, "xml_node"), xml2::xml_name(button) == "button")
+  attr <- as.list(xml2::xml_attrs(button))
 
   structure(
     list(
@@ -216,7 +216,7 @@ format.button <- function(x, ...) {
 #' @return An updated form object
 #' @export
 #' @examples
-#' search <- html_form(html("https://www.google.com"))[[1]]
+#' search <- html_form(read_html("http://www.google.com"))[[1]]
 #' set_values(search, q = "My little pony")
 #' set_values(search, hl = "fr")
 #' \dontrun{set_values(search, btnI = "blah")}
@@ -263,7 +263,7 @@ set_values <- function(form, ...) {
 #' f1 <- set_values(f0, entry.564397473 = "abc")
 submit_form <- function(session, form, submit = NULL, ...) {
   request <- submit_request(form, submit)
-  url <- XML::getRelativeURL(session$url, form$url)
+  url <- xml2::url_relative(form$url, session$url)
 
   # Make request
   if (request$method == "GET") {
@@ -322,5 +322,5 @@ submit_request <- function(form, submit = NULL) {
 #' @examples
 #' google_form("1M9B8DsYNFyDjpwSK6ur_bZf8Rv_04ma3rmaaBiveoUI")
 google_form <- function(x) {
-  html(paste0("https://docs.google.com/forms/d/", x, "/viewform"))
+  xml2::read_html(httr::GET(paste0("https://docs.google.com/forms/d/", x, "/viewform")))
 }