+++ R/xml-methods.R
@@ -1,4 +1,0 @@
-#' @export
-`[.XMLNodeSet` <- function(x, i, ...) {
-  structure(NextMethod(),  class = "XMLNodeSet")
-}